wip: refactor folder structure
