+++ tns-core-modules/ui/bottom-navigation/package.json
@@ -1,5 +1,5 @@
 {
-  "name" : "ui/bottom-navigation",
+  "name" : "bottom-navigation",
   "main" : "bottom-navigation",
   "types" : "bottom-navigation.d.ts"
 }
