+++ tns-core-modules/ui/builder/component-builder/component-builder.ts
@@ -16,9 +16,9 @@ import * as filesystem from "../../../file-system";
 const UI_PATH = "ui/";
 const MODULES = {
     "TabViewItem": "ui/tab-view",
-    "TabStrip": "ui/bottom-navigation",
-    "TabStripItem": "ui/bottom-navigation",
-    "TabContentItem": "ui/bottom-navigation",
+    "TabStrip": "ui/tab-navigation-base/tab-strip",
+    "TabStripItem": "ui/tab-navigation-base/tab-strip-item",
+    "TabContentItem": "ui/tab-navigation-base/tab-content-item",
     "FormattedString": "text/formatted-string",
     "Span": "text/span",
     "ActionItem": "ui/action-bar",
