+++ tns-core-modules/ui/tab-navigation-base/tab-content-item/package.json
