+++ tns-core-modules/ui/tab-navigation-base/tab-content-item/tab-content-item.android.ts
