+++ tns-core-modules/ui/tab-navigation-base/tab-navigation-base/tab-navigation-base.ts
