+++ tns-core-modules/ui/tab-navigation-base/tab-strip-item/package.json
