+++ tns-core-modules/ui/tab-navigation-base/tab-strip-item/tab-strip-item.d.ts
