+++ tns-core-modules/ui/tab-navigation-base/tab-strip/package.json
