+++ tns-core-modules/ui/tab-navigation-base/tab-strip/tab-strip.ts
