--- tns-core-modules/ui/tab-navigation/tabs/package.json
@@ -1,5 +1,0 @@
-{
-  "name" : "tabs",
-  "main" : "tabs",
-  "types" : "tabs.d.ts"
-}
