+++ tns-core-modules/ui/tabs/package.json
@@ -1,5 +1,5 @@
 {
-  "name" : "ui/tabs",
+  "name" : "tabs",
   "main" : "tabs",
   "types" : "tabs.d.ts"
 }
